refactor: 로그인하지 않은 사용자가 특정 페이지 접속 시, ExceptionHandler를 통해 로그인 페이지가 보여지도록 변경
1. UserController에서 UnknownHostAccessException 발생 시 ExceptionHandler로 던져지도록 변경한다
2. UnknownHostAccessException 익셉션 클래스를 추가한다
3. 로그인 뷰가 에러 메시지를 받으면 alert로 출력되도록 수정한다
